requirements.txt
pillow
numpy
pandas
scikit-learn
optuna
tensorboard
matplotlib
seaborn
tqdm
seaborn
medmnist
torchmetrics
mlxtend